@@ -3,47 +3,45 @@
   "version": "0.1.0",
   "private": true,
   "description": "MVP for Newsletter Discovery Platform",
+  "type": "module",
   "scripts": {
-    "start": "set SKIP_PREFLIGHT_CHECK=true && react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
     "lint": "eslint src/**/*.{ts,tsx}",
     "format": "prettier --write src/**/*.{ts,tsx}",
-    "eject": "react-scripts eject"
+    "seed": "node scripts/seed.js",
+    "seed:newsletters": "node scripts/seedNewsletters.js",
+    "seed:user-preferences": "node scripts/seedUserPreferences.js",
+    "seed:user-interactions": "node scripts/seedUserInteractions.js",
+    "seed:all": "npm run seed:newsletters && npm run seed:user-preferences && npm run seed:user-interactions"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.9.5"
+    "firebase": "^9.22.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.11.2",
+    "zustand": "^4.3.8"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.6",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
-    "eslint": "^7.11.0",
-    "prettier": "^2.4.1"
+    "@tailwindcss/forms": "^0.5.3",
+    "@tailwindcss/typography": "^0.5.9",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@vitejs/plugin-react": "^4.0.3",
+    "autoprefixer": "^10.4.14",
+    "eslint": "^8.36.0",
+    "postcss": "^8.4.24",
+    "prettier": "^2.8.4",
+    "tailwindcss": "^3.3.2",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.4",
+    "vite": "^4.3.9"
   },
   "eslintConfig": {
     "extends": [
       "react-app",
       "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
     ]
   }
 }
