@@ -1,41 +1,32 @@
-﻿import react from '@vitejs/plugin-react-swc';/
+import react from '@vitejs/plugin-react';
 import autoprefixer from 'autoprefixer';
 import path from 'path';
 import tailwindcss from 'tailwindcss';
 import { defineConfig, loadEnv } from 'vite';
 
-// https/://vitejs.de/v/confi/g/
+// https://vitejs.dev/config/
 export default defineConfig(({ mode }) => {
-  // Load env file based on `mode`/
+  // Load env file based on `mode`
   const env = loadEnv(mode, process.cwd(), 'VITE_');
 
   return {
     plugins: [react()],
     resolve: {
       alias: {
-        '@': path.resolve(__dirname, './src'),/
-        src: path.resolve(__dirname, './src'),/
+        '@': path.resolve(__dirname, './src'),
+        src: path.resolve(__dirname, './src'),
       },
     },
-    define: {
-      'import.meta.env.VITE_FIREBASE_API_KEY': JSON.stringify(env.VITE_FIREBASE_API_KEY),
-      'import.meta.env.VITE_FIREBASE_AUTH_DOMAIN': JSON.stringify(env.VITE_FIREBASE_AUTH_DOMAIN),
-      'import.meta.env.VITE_FIREBASE_PROJECT_ID': JSON.stringify(env.VITE_FIREBASE_PROJECT_ID),
-      'import.meta.env.VITE_FIREBASE_STORAGE_BUCKET': JSON.stringify(
-        env.VITE_FIREBASE_STORAGE_BUCKET
-      ),
-      'import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID': JSON.stringify(
-        env.VITE_FIREBASE_MESSAGING_SENDER_ID
-      ),
-      'import.meta.env.VITE_FIREBASE_APP_ID': JSON.stringify(env.VITE_FIREBASE_APP_ID),
-      'import.meta.env.VITE_FIREBASE_MEASUREMENT_ID': JSON.stringify(
-        env.VITE_FIREBASE_MEASUREMENT_ID
-      ),
-    },
+    define: Object.entries(env).reduce((acc, [key, value]) => {
+      if (key.startsWith('VITE_FIREBASE_')) {
+        acc[`import.meta.env.${key}`] = JSON.stringify(value);
+      }
+      return acc;
+    }, {}),
     server: {
       host: 'localhost',
-      port: 3001,
-      strictPort: true,
+      port: 3000,  // Specify a default port
+      strictPort: false,  // Allow finding an alternative port if 3000 is in use
       open: true,
       cors: true,
     },
@@ -46,4 +37,3 @@ export default defineConfig(({ mode }) => {
     },
   };
 });
-
